03-setting-up-data-lake-using-google-cloud-storage/pyproject.toml
[tool.poetry]
name = "03-setting-up-data-lake-using-google-cloud-storage"
version = "0.1.0"
description = ""
authors = ["Tor Thanawat <60775527+thanawat-hub@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-storage = "^2.14.0"
sqlalchemy = "^2.0.25"
pymysql = "^1.1.0"
pandas = "^2.1.4"
google-cloud-bigquery = "^3.16.0"
pyarrow = "^14.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
